refactor: move styles from app to styles folder
