+++ src/@types/styled.d.ts
@@ -1,5 +1,5 @@
 import "styled-components/native";
-import type { ThemeType } from "../styles/theme";
+import type { ThemeType } from "../styles/themes/default";
 
 declare module "styled-components/native" {
   export interface DefaultTheme extends ThemeType {}
