+++ src/app/(app)/(create)/index.tsx
@@ -1,5 +1,6 @@
 import { Text } from "react-native";
-import { Container } from "./styles";
+
+import { Container } from "@/styles/app/create/styles";
 
 export default function CreatePage() {
   return (
