+++ src/app/(auth)/_layout.tsx
@@ -1,9 +1,9 @@
 import { Slot } from "expo-router";
+import React from "react";
 
-import { FormContainer, LayoutContainer } from "./styles";
+import { FormContainer, LayoutContainer } from "@/styles/auth/styles";
 
 import authBackground from "@/assets/auth-background.jpg";
-import React from "react";
 
 export default function AuthLayout() {
   return (
