+++ src/app/_layout.tsx
@@ -3,12 +3,12 @@ import {
   LeagueSpartan_700Bold,
   useFonts,
 } from "@expo-google-fonts/league-spartan";
-import { ThemeProvider } from "styled-components/native";
-
-import { defaultTheme } from "@/styles/theme";
 import { Slot } from "expo-router";
 import { ActivityIndicator, View } from "react-native";
 import { GestureHandlerRootView } from "react-native-gesture-handler";
+import { ThemeProvider } from "styled-components/native";
+
+import { defaultTheme } from "@/styles/themes/default";
 
 export default function RootLayout() {
   const [fontsLoaded] = useFonts({
