+++ src/styles/app/memes/trending.styles.ts
