+++ src/styles/auth/forgot-password.styles.ts
