+++ src/styles/auth/register.styles.ts
