+++ src/styles/auth/styles.ts
